+++ src/components/_private/Flyout/Flyout.types.ts
@@ -82,6 +82,7 @@ type BaseProps = {
 	trapFocusMode?: TrapMode;
 	disabled?: boolean;
 	disableHideAnimation?: boolean;
+	disableContentHover?: boolean;
 	children?: React.ReactNode;
 	onOpen?: () => void;
 	onClose?: () => void;
@@ -137,6 +138,7 @@ export type ContextProps = {
 	| "trapFocusMode"
 	| "contentGap"
 	| "containerRef"
+	| "disableContentHover"
 >;
 
 export type TriggerContextProps = Pick<ContextProps, "triggerElRef">;