+++ src/components/_private/Flyout/FlyoutControlled.tsx
@@ -27,6 +27,7 @@ const FlyoutRoot = (props: T.ControlledProps & T.DefaultProps) => {
 		trapFocusMode,
 		width,
 		disableHideAnimation,
+		disableContentHover,
 		contentGap,
 		contentClassName,
 		contentAttributes,
@@ -314,6 +315,7 @@ const FlyoutRoot = (props: T.ControlledProps & T.DefaultProps) => {
 				contentClassName,
 				contentAttributes,
 				containerRef,
+				disableContentHover,
 			}}
 		>
 			{children}